+++ src/super/i-data/CHANGELOG.md
@@ -9,7 +9,7 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-##   v3.0.0-rc.71 ()
+##   v3.0.0-rc.72 ()
 
 #### :boom: Breaking Change